Add account form + entity validation (#32)
The changes are bigger than expected as I added plumbing to have proper validation using class-validator. The good thing about the approach here is that the validation we have for the class applies to the validation in the form so we are re-using the code. In a summarised way, this PR brings:

- A new form to add accounts
- Modified Transaction form to have a better UI and dealing with splits
- Validation that is reused both for inserting objects to the DB and for the forms.
- All entities now have some minimum validation that force us to create them in a specific way (for example: Transaction object always has to be created with splits). This is what has made the biggest changes in this PR as had to change loads of tests to fix the creation of entities to follow the validation rules.